requirements.txt
torch
torchvision
torchaudio
transformers
pillow
streamlit
